pages/users: Rename component from UserModel.component to UserModal
- Also use withStyles to style components in UserModal
- Disable row selection on click in DataGrid since it is of no use to us
