+++ src/pages/users/components/UserModal.jsx
@@ -1,10 +1,22 @@
 import React from 'react';
-import { Modal, Paper, Grid, CircularProgress } from '@material-ui/core';
+import { Modal, Paper, Grid, CircularProgress, withStyles } from '@material-ui/core';
 import { connect } from 'react-redux';
 
 import FormInput from 'components/FormInput/FormInput.component';
 import { FetchUser, CreateUser, UpdateUser } from 'redux/user/action'
 import Button from 'components/Button/Button.component';
+
+const styles = theme => ({
+    modalBody: {
+        top: "50%",
+        left: "50%",
+        position: "absolute",
+        transform: "translate(-50%,-50%)",
+        display: "flex",
+        flexDirection: "column",
+        padding: 50
+    }
+});
 
 class EditModal extends React.Component {
     constructor(props) {
@@ -51,18 +63,12 @@ class EditModal extends React.Component {
     }
 
     render() {
+        const { classes } = this.props;
+
         if (this.props.isLoading) {
             return (
                 <Modal open={this.props.isOpen} onClose={this.props.onClose}>
-                    <Paper style={{
-                        top: "50%",
-                        left: "50%",
-                        position: "absolute",
-                        transform: "translate(-50%,-50%)",
-                        display: "flex",
-                        flexDirection: "column",
-                        padding: 100
-                    }} component={Grid} item>
+                    <Paper className={classes.modalBody} component={Grid} item>
                         <CircularProgress size="3em" />
                     </Paper>
                 </Modal>
@@ -71,14 +77,7 @@ class EditModal extends React.Component {
 
         return (
             <Modal open={this.props.isOpen} onClose={this.props.onClose}>
-                <Paper style={{
-                    top: "50%",
-                    left: "50%",
-                    position: "absolute",
-                    transform: "translate(-50%,-50%)",
-                    display: "flex",
-                    flexDirection: "column"
-                }} component={Grid} item>
+                <Paper className={classes.modalBody} component={Grid} item>
                     <FormInput type="text" name="full_name" placeholder="Full Name" value={this.state.user.full_name ?? ""} onChange={this.handleInputChange} />
                     <FormInput type="email" name="email" placeholder="Email" value={this.state.user.email ?? ""} onChange={this.handleInputChange} />
                     <FormInput type="checkbox" name="is_active" color="primary" checked={this.state.user.is_active ?? false} onChange={this.handleInputChange} />
@@ -99,4 +98,6 @@ const mapStateToProps = (state) => ({
     isLoading: state.user.isLoading
 })
 
-export default connect(mapStateToProps, { FetchUser, CreateUser, UpdateUser })(EditModal);
+export default withStyles(styles)(
+    connect(mapStateToProps, { FetchUser, CreateUser, UpdateUser })(EditModal)
+);
